+++ ConsentScreens/Config.ts
@@ -1,0 +1,5 @@
+export const config = {
+    TERMS_URL: 'https://cdn.websitepolicies.com/wp-content/uploads/2017/06/terms-and-conditions-template-1038x1536.png',
+    SECURE_URL: 'https://www.docformats.com/wp-content/uploads/2023/11/Pre-Order-Terms-and-Conditions-Template-Word-Google-Docs.png',
+    
+};
