+++ package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@react-native-community/checkbox": "^0.5.17",
     "react": "18.2.0",
-    "react-native": "0.74.1"
+    "react-native": "0.74.1",
+    "react-native-safe-area-context": "^4.10.5"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
